+++ x/configuration/types/fees_test.go
@@ -1,19 +1,20 @@
 package types
 
 import (
-	"github.com/cosmos/cosmos-sdk/types"
 	"testing"
+
+	"github.com/cosmos/cosmos-sdk/types"
 )
 
 func TestFees_Validate(t *testing.T) {
 	type fields struct {
 		FeeCoinDenom                 string
 		FeeCoinPrice                 types.Dec
-		DefaultFee                   types.Dec
-		RegisterClosedAccount        types.Dec
-		RegisterOpenAccount          types.Dec
-		TransferClosedAccount        types.Dec
-		TransferOpenAccount          types.Dec
+		FeeDefault                   types.Dec
+		RegisterAccountClosed        types.Dec
+		RegisterAccountOpen          types.Dec
+		TransferAccountClosed        types.Dec
+		TransferAccountOpen          types.Dec
 		ReplaceAccountTargets        types.Dec
 		AddAccountCertificate        types.Dec
 		DelAccountCertificate        types.Dec
@@ -25,9 +26,9 @@ func TestFees_Validate(t *testing.T) {
 		RegisterDomain5              types.Dec
 		RegisterDomainDefault        types.Dec
 		RegisterOpenDomainMultiplier types.Dec
-		TransferClosedDomain         types.Dec
-		TransferOpenDomain           types.Dec
-		RenewOpenDomain              types.Dec
+		TransferDomainClosed         types.Dec
+		TransferDomainOpen           types.Dec
+		RenewDomainOpen              types.Dec
 	}
 	tests := []struct {
 		name    string
@@ -54,11 +55,11 @@ func TestFees_Validate(t *testing.T) {
 			f := &Fees{
 				FeeCoinDenom:                 tt.fields.FeeCoinDenom,
 				FeeCoinPrice:                 tt.fields.FeeCoinPrice,
-				DefaultFee:                   tt.fields.DefaultFee,
-				RegisterClosedAccount:        tt.fields.RegisterClosedAccount,
-				RegisterOpenAccount:          tt.fields.RegisterOpenAccount,
-				TransferClosedAccount:        tt.fields.TransferClosedAccount,
-				TransferOpenAccount:          tt.fields.TransferOpenAccount,
+				FeeDefault:                   tt.fields.FeeDefault,
+				RegisterAccountClosed:        tt.fields.RegisterAccountClosed,
+				RegisterAccountOpen:          tt.fields.RegisterAccountOpen,
+				TransferAccountClosed:        tt.fields.TransferAccountClosed,
+				TransferAccountOpen:          tt.fields.TransferAccountOpen,
 				ReplaceAccountTargets:        tt.fields.ReplaceAccountTargets,
 				AddAccountCertificate:        tt.fields.AddAccountCertificate,
 				DelAccountCertificate:        tt.fields.DelAccountCertificate,
@@ -70,9 +71,9 @@ func TestFees_Validate(t *testing.T) {
 				RegisterDomain5:              tt.fields.RegisterDomain5,
 				RegisterDomainDefault:        tt.fields.RegisterDomainDefault,
 				RegisterOpenDomainMultiplier: tt.fields.RegisterOpenDomainMultiplier,
-				TransferClosedDomain:         tt.fields.TransferClosedDomain,
-				TransferOpenDomain:           tt.fields.TransferOpenDomain,
-				RenewOpenDomain:              tt.fields.RenewOpenDomain,
+				TransferDomainClosed:         tt.fields.TransferDomainClosed,
+				TransferDomainOpen:           tt.fields.TransferDomainOpen,
+				RenewDomainOpen:              tt.fields.RenewDomainOpen,
 			}
 			if err := f.Validate(); (err != nil) != tt.wantErr {
 				t.Errorf("Validate() error = %v, wantErr %v", err, tt.wantErr)